services/telematics_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
trino==0.330.0
pytz==2024.1
PyYAML==6.0.2